# swagger.yaml
basePath: /v1/connectionmgmt
consumes:
    - application/json
definitions:
    AWSConnection:
        description: AWSConnection represents AWSConnection resource serialized by Microservice endpoints
        x-go-package: DemoServer_APPLICATIONMANAGER/data
...
    ErrorResponse:
        description: |-
            ErrorResponse represents information returned by Microservice endpoints in case that was an error
            in normal execution flow.
        x-go-package: DemoServer_APPLICATIONMANAGER/helper
...
info:
    contact:
        email: muhammad.w.ahmed.48@gmail.com
        name: Muhammad Ahmed
    description: |-
        Connection defines cloud account connection for applicaiton deployment, spin up and down.
        APPLICATIONMANAGER serves as one stop shop for CRUD as well as connectivity testing capabilities for all connection types for DemoServer.
    termsOfService: there are no TOS at this moment, use at your own risk we take no responsibility
    title: DemoServer_APPLICATIONMANAGER API.
    version: 1.0.0
paths:
...
    /connections:
        get:
            description: |-
                Endpoint: GET - /v1/connectionmgmt/connections

                Description: Returns list of generic connections resources. It is useful to list all connections
                currently in APPLICATIONMANAGER. Generic Connection resource does not have specific details
                and attributes of specialized connection types. It only tracks general information about
                connection including its type.
            operationId: GetConnections
            parameters:
                - description: maximum number of results to return.
                  format: int32
                  in: query
                  name: limit
                  type: integer
                - description: number of results to be skipped from beginning of list
                  format: int32
                  in: query
                  name: skip
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: List of AWSConnection resources
                    schema:
                        items:
                            $ref: '#/definitions/AWSConnection'
                        type: array
                "400":
                    description: Issues with parameters or their value
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                "500":
                    description: Internal server error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                default:
                    description: unexpected error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            summary: List Connections
            tags:
                - Connection
...
produces:
    - application/json
schemes:
    - http
swagger: "2.0"